Week4_Responsive_Web_design
My responsive web design is based on a Korean card game, Go-Stop. Hwatu
is a deck of cards used to play a number of games. The most common one
is Go-Stop, which is Korea's most beloved card game. This card game is
similar to a poker game. The reason why I am designing Go-Stop game is
because I have always wanted to put something transitional onto my
design work. So I decided to design a web design with Korea's
traditional art, Hwatu.